:white_check_mark: Add tests proving that get current fiscal function is actually correct.
